requirements.txt
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
distlib==0.3.8
filelock==3.15.4
idna==3.8
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
platformdirs==4.2.2
pycparser==2.22
pypdfium2==4.30.0
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
urllib3==2.2.2
virtualenv==20.26.3
